Collector/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2
djangorestframework==3.16.0
dnspython==2.7.0
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipwhois==1.3.0
ldap3==2.9.1
multidict==6.2.0
numpy==2.2.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
propcache==0.3.1
pyasn1==0.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nmap==0.7.1
python-whois==0.9.5
pytz==2025.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whois==1.20240129.2
wsproto==1.2.0
yarl==1.19.0
